Move examples into pyomo/examples/gdp
