--- examples/gdp/ConstrainedLayout_BigM.dat
@@ -1,35 +1,0 @@
-set BOXES := 1 2 3 ;
-set CIRCLES := 1 2 ;
-
-param Length :=
-    1  5
-    2  7
-    3  3
-;
-
-param Height :=
-    1  6
-    2  5
-    3  3
-;
-
-param CenterXCoord :=
-    1  15
-    2  50
-;
-
-param CenterYCoord :=
-    1  10
-    2  80
-;
-
-param Radius :=
-    1  6
-    2  5
-;
-
-param DistCost :=
-    1  2  300
-    1  3  240
-    2  3  100
-;
