--- examples/gdp/stripPacking_12Chull.dat
@@ -1,34 +1,0 @@
-set RECTANGLES := 1 2 3 4 5 6 7 8 9 10 11 12 ;
-
-param Heights :=
-    1 10
-    2 9
-    3 8
-    4 4
-    5 5
-    6 6
-    7 7
-    8 3
-    9 2
-    10 1
-    11 1
-    12 3
-;
-
-param Lengths :=
-    1 1
-    2 2
-    3 3
-    4 4
-    5 5
-    6 9
-    7 7
-    8 6
-    9 5
-    10 12
-    11 3
-    12 2
-;
-
-param StripWidth := 10 ;
-param LengthUB := 27 ;
