+++ examples/gdp/strip_packing/stripPacking_4BigM.dat
@@ -1,0 +1,18 @@
+set RECTANGLES := 1 2 3 4 ;
+
+param Heights :=
+    1 6
+    2 3
+    3 4
+    4 2
+;
+
+param Lengths :=
+    1 6
+    2 8
+    3 5
+    4 3
+;
+
+param StripWidth := 10 ;
+param LengthUB := 15 ;
