+++ examples/gdp/strip_packing/strip_packing_concrete.py
